Improve concurrent rendering
Distribute pixels to render randomly between available threads in order
to make each of them work on average the same amount of time.

Previously there was a possibility that one ore more threads got
easier part of the image to render and returned significantly earlier
than others.

"I paid for the whole CPU, so I'm gonna use the whole CPU."